Address feedback move logic to a lib file. Remove mentions of hardcoded paths.
